*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@6933 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/comm.c
@@ -1348,7 +1348,7 @@ bool writer( int type, ... )
 				return pipe_write( ( char * ) &header, sizeof header );
 
 			case C_FLOAT :
-				header.data.float_data = va_arg( ap, float );
+				header.data.float_data = ( float ) va_arg( ap, double );
 				va_end( ap );
 				return pipe_write( ( char * ) &header, sizeof header );
